+++ lib/dradis/plugins/csv/mapping.rb
@@ -1,0 +1,6 @@
+module Dradis::Plugins::CSV
+  module Mapping
+    DEFAULT_MAPPING = {}.freeze
+    SOURCE_FIELDS = {}.freeze
+  end
+end